requirements.txt
pytest
selene>=2.0.0rc6
allure-pytest==2.13.5
